implementations/python/pyproject.toml
[project]
name = "reifinator"
version = "0.1.0"
description = "Template-structure-driven code and directory generator"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mako = ["mako>=1.2.0"]
jinja2 = ["jinja2>=3.1.0"]
all = ["mako>=1.2.0", "jinja2>=3.1.0"]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
reify = "reifinator.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
